requirements.txt
alembic==0.9.6
amqp==2.2.2
appnope==0.1.0
astroid==1.5.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Celery-Helper==1.1.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
kombu==4.1.0
lazy-object-proxy==1.3.1
ldap3==2.4
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
Pygments==2.2.0
pylint==1.7.4
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
python-gitlab==1.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.15
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.12.2
wrapt==1.10.11
WTForms==2.1
